pyproject.toml
[tool.poetry]
name = "meetshort"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
flask = "^3.0.3"
helper = "^2.5.0"
werkzeug = "^3.0.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
gtts = "^2.5.1"
langchain-core = "^0.1.40"
langchain-google-genai = "^1.0.1"
moviepy = "^1.0.3"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"